requirements.txt
pytesseract
Pillow
opencv-python
python-pptx
faiss-cpu
requests
easyocr
ollama
streamlit